add WDT capability to update operation
