+++ imagetool/src/main/resources/docker-files/Update_Image.mustache
@@ -4,14 +4,114 @@
 # Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
 #
 #
+{{#isWdtEnabled}}
+    FROM {{baseImage}} as WDT_BUILD
+    LABEL WlsImageToolLayer="intermediate"
+
+    ARG WDT_PKG
+    ARG DOMAIN_TYPE
+    ARG DOMAIN_PARENT
+    ARG DOMAIN_HOME
+    ARG WDT_ARCHIVE
+    ARG WDT_VARIABLE
+    ARG ADMIN_NAME
+    ARG ADMIN_HOST
+    ARG ADMIN_PORT
+    ARG MANAGED_SERVER_PORT
+    ARG SCRIPTS_DIR
+    ARG WDT_HOME
+    ARG RCU_RUN_FLAG
+
+    ENV WDT_PKG=${WDT_PKG:-weblogic-deploy.zip} \
+    ADMIN_NAME=${ADMIN_NAME:-admin-server} \
+    ADMIN_HOST=${ADMIN_HOST:-wlsadmin} \
+    ADMIN_PORT=${ADMIN_PORT:-7001} \
+    JAVA_HOME={{{java_home}}} \
+    MANAGED_SERVER_NAME=${MANAGED_SERVER_NAME:-} \
+    MANAGED_SERVER_PORT=${MANAGED_SERVER_PORT:-8001} \
+    WLSDEPLOY_PROPERTIES="-Djava.security.egd=file:/dev/./urandom" \
+    DOMAIN_TYPE=${DOMAIN_TYPE:-WLS} \
+    DOMAIN_PARENT=${DOMAIN_PARENT:-/u01/domains} \
+    WDT_ARCHIVE=${WDT_ARCHIVE:-} \
+    WDT_VARIABLE=${WDT_VARIABLE:-} \
+    LC_ALL=${DEFAULT_LOCALE:-en_US.UTF-8} \
+    PROPERTIES_FILE_DIR={{{oracle_home}}}/properties \
+    WDT_HOME=${WDT_HOME:-/u01/app/weblogic-deploy} \
+    SCRIPTS_DIR=${SCRIPTS_DIR:-scripts} \
+    RCU_RUN_FLAG=${RCU_RUN_FLAG:-} \
+    DOMAIN_HOME=${DOMAIN_HOME:-/u01/domains/base_domain} \
+    PATH=$PATH:{{{java_home}}}/bin:{{{oracle_home}}}/oracle_common/common/bin:{{{oracle_home}}}/wlserver/common/bin:${DOMAIN_HOME}/bin:{{{oracle_home}}}
+
+    USER root
+
+    {{#useYum}}
+        RUN yum -y --downloaddir={{{tmpDir}}} install gzip tar unzip \
+        && yum -y --downloaddir={{{tmpDir}}} clean all \
+        && rm -rf {{{tmpDir}}}
+    {{/useYum}}
+    {{#useAptGet}}
+        RUN apt-get -y update \
+        && apt-get -y upgrade \
+        && apt-get -y install gzip tar unzip \
+        && apt-get -y clean all
+    {{/useAptGet}}
+    {{#useApk}}
+        RUN apk update \
+        && apk upgrade \
+        && rm -rf /var/cache/apk/*
+    {{/useApk}}
+    {{#useZypper}}
+        RUN zypper -nq update \
+        && zypper -nq clean \
+        && rm -rf /var/cache/zypp/*
+    {{/useZypper}}
+
+    COPY --chown={{userid}}:{{groupid}} ${WDT_PKG} {{#wdtModels}}{{.}} {{/wdtModels}}${WDT_ARCHIVE} ${WDT_VARIABLE} {{{tmpDir}}}/
+
+    USER {{userid}}
+
+    RUN unzip -q {{{tmpDir}}}/$WDT_PKG -d $(dirname $WDT_HOME) \
+    && mkdir -p $(dirname ${DOMAIN_HOME}) \
+    && mkdir -p ${PROPERTIES_FILE_DIR} \
+    && if [ -n "$WDT_ARCHIVE" ]; then ARCHIVE_OPT="-archive_file {{{tmpDir}}}/${WDT_ARCHIVE##*/}"; fi \
+    && if [ -n "$WDT_VARIABLE" ]; then VARIABLE_OPT="-variable_file {{{tmpDir}}}/${WDT_VARIABLE##*/}"; fi \
+    && if [ -n "${RCU_RUN_FLAG}" ]; then RCU_RUN_OPT="-run_rcu"; fi \
+    && cd ${WDT_HOME}/bin \
+    && ${WDT_HOME}/bin/createDomain.sh \
+    -oracle_home {{{oracle_home}}} \
+    -java_home {{{java_home}}} \
+    -domain_home ${DOMAIN_HOME} \
+    -domain_type ${DOMAIN_TYPE} \
+    $RCU_RUN_OPT \
+    $VARIABLE_OPT \
+    {{{wdtModelFileArgument}}} \
+    $ARCHIVE_OPT \
+    && rm -rf ${WDT_HOME} {{{tempDir}}}
+{{/isWdtEnabled}}
 
 FROM {{baseImage}} as OS_UPDATE
 USER root
 
 ARG PATCHDIR
+{{#isWdtEnabled}}
+ARG DOMAIN_PARENT
+ARG DOMAIN_HOME
+{{/isWdtEnabled}}
 
 ENV PATCHDIR=${PATCHDIR:-patches} \
     OPATCH_NO_FUSER=true
+
+{{#isWdtEnabled}}
+ENV ADMIN_NAME=${ADMIN_NAME:-admin-server} \
+    ADMIN_HOST=${ADMIN_HOST:-wlsadmin} \
+    ADMIN_PORT=${ADMIN_PORT:-7001} \
+    MANAGED_SERVER_NAME=${MANAGED_SERVER_NAME:-} \
+    MANAGED_SERVER_PORT=${MANAGED_SERVER_PORT:-8001} \
+    WLSDEPLOY_PROPERTIES="-Djava.security.egd=file:/dev/./urandom" \
+    DOMAIN_PARENT=${DOMAIN_PARENT:-/u01/domains} \
+    DOMAIN_HOME=${DOMAIN_HOME:-/u01/domains/base_domain}
+    PATH=$PATH:${DOMAIN_HOME}/bin
+{{/isWdtEnabled}}
 
 USER {{userid}}
 
@@ -30,3 +130,7 @@ RUN $ORACLE_HOME/OPatch/opatch napply -silent -oh $ORACLE_HOME -phBaseDir {{{tmp
  && $ORACLE_HOME/OPatch/opatch util cleanup -silent -oh $ORACLE_HOME \
  && rm -rf {{{tmpDir}}}
 {{/isPatchingEnabled}}
+
+{{#isWdtEnabled}}
+    COPY --from=WDT_BUILD --chown={{userid}}:{{groupid}} $DOMAIN_HOME $DOMAIN_HOME/
+{{/isWdtEnabled}}
